add-show-spec

+++ db/migrate/20210219065801_add_column_to_parkings.rb
@@ -1,5 +1,5 @@
 class AddColumnToParkings < ActiveRecord::Migration[6.1]
   def change
-    add_column :parkings, :price, :bigint
+    add_column :parkings, :price, :integer
   end
 end
